requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-groq>=0.0.3
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
streamlit>=1.32.0
unstructured>=0.11.0
python-docx>=1.0.0
openpyxl>=3.1.2
pandas>=2.0.0
pypdf>=3.17.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'